#9999 refactor az to azure
